+++ just_audio/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.9.38
+
+* Migrate to package:web.
+* Add AudioPlayer.setWebCrossOrigin for CORS on web (@danielwinkler).
+
+## 0.9.37
+
+* Support useLazyPreparation on iOS/macOS.
+* Add index in sequence to errors for Android/iOS/macOS.
+* Fix seek to index UI update on iOS/macOS.
+
 ## 0.9.36
 
 * Add setAllowsExternalPlayback on iOS/macOS.